Renames logger to Logger
