+++ api/src/route/stash.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import passport from "passport";
-import { Logger } from "../../../lib/logger";
+import { Logger } from "../../../lib/Logger";
 import { DataSource } from "typeorm";
 import { User } from "../../../model/User";
 import { Stash } from "../../../model/Stash";
