+++ api/src/route/user.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import passport from "passport";
 import jwt from "jsonwebtoken";
-import { Logger } from "../../../lib/logger";
+import { Logger } from "../../../lib/Logger";
 import { DataSource } from "typeorm";
 import { User } from "../../../model/User";
 import Translations from "../../../lib/Translations";
